add ps2 module and refactor structure
