+++ src/interrupts.rs
@@ -1,4 +1,4 @@
-use crate::{gdt, print, println, scancode};
+use crate::{gdt, println, ps2};
 use core::u8;
 use lazy_static::lazy_static;
 use pic8259::ChainedPics;
@@ -13,11 +13,11 @@ pub enum InterruptIndex {
 }
 
 impl InterruptIndex {
-    fn as_u8(self) -> u8 {
+    pub fn as_u8(self) -> u8 {
         self as u8
     }
 
-    fn as_usize(self) -> usize {
+    pub fn as_usize(self) -> usize {
         usize::from(self.as_u8())
     }
 }
@@ -38,7 +38,7 @@ lazy_static! {
         }
         idt.breakpoint.set_handler_fn(breakpoint_handler);
         idt[InterruptIndex::Timer.as_usize()].set_handler_fn(timer_interrupt_handler);
-        idt[InterruptIndex::Keyboard.as_usize()].set_handler_fn(keyboard_interrupt_handler);
+        idt[InterruptIndex::Keyboard.as_usize()].set_handler_fn(ps2::keyboard_interrupt_handler);
 
         idt
     };
@@ -69,25 +69,5 @@ extern "x86-interrupt" fn timer_interrupt_handler(_stack_frame: InterruptStackFr
     unsafe {
         PICS.lock()
             .notify_end_of_interrupt(InterruptIndex::Timer.as_u8());
-    }
-}
-
-extern "x86-interrupt" fn keyboard_interrupt_handler(_stack_frame: InterruptStackFrame) {
-    use x86_64::instructions::port::Port;
-
-    let mut port = Port::new(0x60);
-    let scancode: u8 = unsafe { port.read() };
-    match scancode::interpret(scancode) {
-        Ok(interp) => {
-            if interp.0 {
-                print!("{}", interp.1);
-            }
-        }
-        Err(masked_scancode) => println!("failed to interpet {}", masked_scancode),
-    }
-
-    unsafe {
-        PICS.lock()
-            .notify_end_of_interrupt(InterruptIndex::Keyboard.as_u8());
     }
 }
